+++ pattern/lib/app/modules/home/home_module.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'views/home/home_controller.dart';
 import 'views/home/home_page.dart';
+
 class HomeModule extends Module {
   @override
   final List<Bind> binds = [