+++ pattern/template/app/modules/template/repositories/template_repository.dart
@@ -1,9 +1,9 @@
 class TemplateRepository {
-  TemplateRepository(this._api, this._box);
+  // TemplateRepository(this._api, this._box);
   static const key = (template: 'template');
   
   //Dependencies
-  final IApi _api;
-  final IBox _box;
+  // final IApi _api;
+  // final IBox _box;
 
 }