+++ src/install.js
@@ -6,18 +6,27 @@ const utils = require('./utils.js')
 var cp = require('child_process')
 
 async function install() {
-    var pubspecPath = await utils.getPubspecPath()
+    var [pubspecPath, success] = await utils.getPubspecPath();
 
     if (typeof pubspecPath === 'string' && pubspecPath.length > 0) {
 
-        /// path = .../
+        if (success) {
+            const overwrite = await vscode.window.showWarningMessage(
+                'This will overwrite everything. Are you sure you want to continue?',
+                { modal: true },
+                'Yes 🔥', 'No'
+            );
+
+            if (overwrite !== 'Yes 🔥') return;
+        }
+
         let path = pubspecPath.replace("pubspec.yaml", "")
         var data = fs.readFileSync(pubspecPath, 'utf-8')
         var lines = data.split('\n')
 
         cp.exec(`
         cd ${path} &&
-        flutter pub remove get flutter_spinkit responsive_framework google_fonts flutter_datetime_picker
+        flutter pub add asp flutter_modular:6.0.0-beta.2
         `, (err, stdout, stderr) => {
             if (err) {
 
@@ -35,7 +44,7 @@ async function install() {
                     index = i
                 }
             }
-            lines.splice(index - 1, 0, "  flutter_modular: 2.2.0")
+            // lines.splice(index - 1, 0, "  flutter_modular: 2.2.0")
             fs.writeFileSync(pubspecPath, lines.join('\n'), 'utf-8')
             data = lines.join('\n')
 
@@ -45,7 +54,8 @@ async function install() {
 
         var projectName = lines[0].replace("name: ", "")
         await moveFile(path, projectName)
-        vscode.window.showInformationMessage('Generate successful 🥳🤘🏻')
+        vscode.window.showInformationMessage('Generate successful 🟨⬛');
+        vscode.window.showInformationMessage('Press F5 to get started 🔥');
     }
     else { return }
 }
@@ -63,22 +73,13 @@ async function moveFile(path, projectName) {
                 replace.sync({
                     files: [
                         `${path}lib/*.dart`,
-                        `${path}lib/app/data/api/*.dart`,
-                        `${path}lib/app/data/provider/*.dart`,
-                        `${path}lib/app/modules/home_module/*.dart`,
-                        `${path}lib/app/modules/splash_module/*.dart`,
-                        `${path}lib/app/routes/*.dart`,
-                        `${path}lib/app/themes/*.dart`,
-                        `${path}lib/app/translations/*.dart`,
-                        `${path}lib/app/utils/*.dart`,
-                        `${path}lib/app/utils/widgets/*.dart`,
-                        `${path}lib/app/utils/widgets/app_bar/*.dart`,
-                        `${path}lib/app/utils/widgets/app_button/*.dart`,
-                        `${path}lib/app/utils/widgets/app_divider/*.dart`,
-                        `${path}lib/app/utils/widgets/app_text_field/*.dart`,
-                        `${path}lib/app/utils/widgets/bottom_sheet_provider/*.dart`,
-                        `${path}lib/app/utils/widgets/dialog_provider/*.dart`,
-                        `${path}lib/app/utils/widgets/dialog_provider/view_dialog/*.dart`,
+                        `${path}lib/app/*.dart`,
+                        `${path}lib/app/modules/*.dart`,
+                        `${path}lib/app/modules/views/*.dart`,
+                        `${path}lib/app/modules/views/home/*.dart`,
+                        `${path}lib/app/modules/services/*.dart`,
+                        `${path}lib/app/modules/repositories/*.dart`,
+                        `${path}lib/app/modules/models/*.dart`,
                     ],
                     from: /modular_generator/g,
                     to: projectName,