+++ src/utils.js
@@ -1,24 +1,26 @@
 const vscode = require('vscode')
 
+/// Returns pubscpec.yaml path and if it came from success or not.
 async function getPubspecPath() {
     try {
         let path = await vscode.workspace.findFiles('pubspec.yaml')
-        return path[0].path
+        return [path[0].path, true];
     } catch (error) {
-        const pickItems = getxInstallTemplate()
+        const pickItems = installModularTemplate()
 
         const selectedTemplate = await vscode.window.showQuickPick(
             pickItems,
             {
                 matchOnDescription: true,
-                placeHolder: "Project not found!",
+                placeHolder: "No pubspec.yaml found. What would you like to do?",
             },
         )
 
         if (!selectedTemplate)
             return;
 
-        chooseGetxInstallTemplate(selectedTemplate.template.id);
+        const path = await chooseInstallModularTemplate(selectedTemplate.template.id);
+        if (typeof path === 'string' && path.length > 0) return [path, false];
         return;
     }
 }
@@ -26,10 +28,10 @@ async function getPubspecPath() {
 /**
  * @param {string} id
  */
-function chooseGetxInstallTemplate(id) {
+async function chooseInstallModularTemplate(id) {
     switch (id) {
         case 'createProject':
-            return createProjectInCurrentFolder();
+            return await createProjectInCurrentFolder();
         case 'openProject':
             return vscode.commands.executeCommand('vscode.openFolder');
         default:
@@ -37,23 +39,38 @@ function chooseGetxInstallTemplate(id) {
     }
 }
 
-function createProjectInCurrentFolder() {
+async function createProjectInCurrentFolder() {
     const terminal = vscode.window.createTerminal('Flutter Create');
-    // const projectName = 'your_project_name'; // Replace this with the desired project name.
-    terminal.sendText(`flutter create .`);
+    terminal.sendText(`flutter create . --empty --overwrite`);
     terminal.show();
+
+    // Timeout after 15 seconds
+    const maxTries = 30;
+    for (let i = 0; i < maxTries; i++) {
+        await new Promise(resolve => setTimeout(resolve, 500)); // 500 ms delay, adjust as necessary
+
+        let pubspecPath = await vscode.workspace.findFiles('pubspec.yaml');
+        if (pubspecPath.length > 0) {
+            console.log('pubspec.yaml has been created');
+            return pubspecPath[0].path;
+        }
+    }
+
+    console.log('pubspec.yaml not found after waiting');
 }
 
-function getxInstallTemplate() {
+
+
+function installModularTemplate() {
     const templates = [
         {
-            detail: "Create new flutter project.",
-            label: "Flutter: New Project",
+            detail: "Install a new Modular project. (project folder must be in snake_case)",
+            label: "Clean install here (overwrite)",
             template: { id: "createProject" },
         },
         {
-            detail: "Open an existing flutter project.",
-            label: "Flutter: Open Project",
+            detail: "Pick a project with a valid pubspec.yaml and call install again.",
+            label: "Open existing Flutter Project",
             template: { id: "openProject" },
         },
     ]